partial migration of cpu, reorganize data file structure
