+++ Cargo.toml
@@ -91,7 +91,7 @@ regex = "1.11.1"
 serde = { version = "1.0.217", features = ["derive"] }
 starship-battery = { version = "0.10.0", optional = true }
 sysinfo = "=0.30.13"
-timeless = "0.0.10-alpha"
+timeless = "0.0.12-alpha"
 toml_edit = { version = "0.22.22", features = ["serde"] }
 tui = { version = "0.29.0", package = "ratatui" }
 unicode-ellipsis = "0.3.0"
