+++ src/app/data/mod.rs
@@ -1,0 +1,10 @@
+//! In charge of cleaning, processing, and managing data.
+
+mod time_series;
+pub use time_series::{TimeSeriesData, Values};
+
+mod process;
+pub use process::ProcessData;
+
+mod store;
+pub use store::*;
