+++ src/canvas/widgets/cpu_basic.rs
@@ -21,113 +21,108 @@ impl Painter {
     pub fn draw_basic_cpu(
         &self, f: &mut Frame<'_>, app_state: &mut App, mut draw_loc: Rect, widget_id: u64,
     ) {
-        // Skip the first element, it's the "all" element
-        if app_state.converted_data.cpu_data.len() > 1 {
-            let cpu_data: &[CpuWidgetData] = &app_state.converted_data.cpu_data[1..];
+        let cpu_data: &[CpuWidgetData] = &app_state.converted_data.cpu_data[1..];
 
-            // This is a bit complicated, but basically, we want to draw SOME number
-            // of columns to draw all CPUs.  Ideally, as well, we want to not have
-            // to ever scroll.
-            // **General logic** - count number of elements in cpu_data.  Then see how
-            // many rows and columns we have in draw_loc (-2 on both sides for border?).
-            // I think what we can do is try to fit in as many in one column as possible.
-            // If not, then add a new column.
-            // Then, from this, split the row space across ALL columns.  From there,
-            // generate the desired lengths.
+        // This is a bit complicated, but basically, we want to draw SOME number
+        // of columns to draw all CPUs.  Ideally, as well, we want to not have
+        // to ever scroll.
+        // **General logic** - count number of elements in cpu_data.  Then see how
+        // many rows and columns we have in draw_loc (-2 on both sides for border?).
+        // I think what we can do is try to fit in as many in one column as possible.
+        // If not, then add a new column.
+        // Then, from this, split the row space across ALL columns.  From there,
+        // generate the desired lengths.
 
-            if app_state.current_widget.widget_id == widget_id {
-                f.render_widget(
-                    widget_block(true, true, self.styles.border_type)
-                        .border_style(self.styles.highlighted_border_style),
-                    draw_loc,
-                );
-            }
+        if app_state.current_widget.widget_id == widget_id {
+            f.render_widget(
+                widget_block(true, true, self.styles.border_type)
+                    .border_style(self.styles.highlighted_border_style),
+                draw_loc,
+            );
+        }
 
-            let (cpu_data, avg_data) =
-                maybe_split_avg(cpu_data, app_state.app_config_fields.dedicated_average_row);
+        let (cpu_data, avg_data) =
+            maybe_split_avg(cpu_data, app_state.app_config_fields.dedicated_average_row);
 
-            if let Some(avg) = avg_data {
-                let (outer, inner, ratio, style) = self.cpu_info(&avg);
-                let [cores_loc, mut avg_loc] =
-                    Layout::vertical([Constraint::Min(0), Constraint::Length(1)]).areas(draw_loc);
+        if let Some(avg) = avg_data {
+            let (outer, inner, ratio, style) = self.cpu_info(&avg);
+            let [cores_loc, mut avg_loc] =
+                Layout::vertical([Constraint::Min(0), Constraint::Length(1)]).areas(draw_loc);
 
-                // The cores section all have horizontal margin, so to line up with the cores we
-                // need to add some margin ourselves.
-                avg_loc.x += 1;
-                avg_loc.width -= 2;
+            // The cores section all have horizontal margin, so to line up with the cores we
+            // need to add some margin ourselves.
+            avg_loc.x += 1;
+            avg_loc.width -= 2;
 
-                f.render_widget(
-                    PipeGauge::default()
-                        .gauge_style(style)
-                        .label_style(style)
-                        .inner_label(inner)
-                        .start_label(outer)
-                        .ratio(ratio),
-                    avg_loc,
-                );
+            f.render_widget(
+                PipeGauge::default()
+                    .gauge_style(style)
+                    .label_style(style)
+                    .inner_label(inner)
+                    .start_label(outer)
+                    .ratio(ratio),
+                avg_loc,
+            );
 
-                draw_loc = cores_loc;
-            }
+            draw_loc = cores_loc;
+        }
 
-            if draw_loc.height > 0 {
-                let remaining_height = usize::from(draw_loc.height);
-                const REQUIRED_COLUMNS: usize = 4;
+        if draw_loc.height > 0 {
+            let remaining_height = usize::from(draw_loc.height);
+            const REQUIRED_COLUMNS: usize = 4;
 
-                let col_constraints =
-                    vec![Constraint::Percentage((100 / REQUIRED_COLUMNS) as u16); REQUIRED_COLUMNS];
-                let columns = Layout::default()
-                    .constraints(col_constraints)
-                    .direction(Direction::Horizontal)
-                    .split(draw_loc);
+            let col_constraints =
+                vec![Constraint::Percentage((100 / REQUIRED_COLUMNS) as u16); REQUIRED_COLUMNS];
+            let columns = Layout::default()
+                .constraints(col_constraints)
+                .direction(Direction::Horizontal)
+                .split(draw_loc);
 
-                let mut gauge_info = cpu_data.iter().map(|cpu| self.cpu_info(cpu));
+            let mut gauge_info = cpu_data.iter().map(|cpu| self.cpu_info(cpu));
 
-                // Very ugly way to sync the gauge limit across all gauges.
-                let hide_parts = columns
-                    .first()
-                    .map(|col| {
-                        if col.width >= 12 {
-                            LabelLimit::None
-                        } else if col.width >= 10 {
-                            LabelLimit::Bars
-                        } else {
-                            LabelLimit::StartLabel
-                        }
-                    })
-                    .unwrap_or_default();
+            // Very ugly way to sync the gauge limit across all gauges.
+            let hide_parts = columns
+                .first()
+                .map(|col| {
+                    if col.width >= 12 {
+                        LabelLimit::None
+                    } else if col.width >= 10 {
+                        LabelLimit::Bars
+                    } else {
+                        LabelLimit::StartLabel
+                    }
+                })
+                .unwrap_or_default();
 
-                let num_entries = cpu_data.len();
-                let mut row_counter = num_entries;
-                for (itx, column) in columns.iter().enumerate() {
-                    if REQUIRED_COLUMNS > itx {
-                        let to_divide = REQUIRED_COLUMNS - itx;
-                        let num_taken = min(
-                            remaining_height,
-                            (row_counter / to_divide) + usize::from(row_counter % to_divide != 0),
+            let num_entries = cpu_data.len();
+            let mut row_counter = num_entries;
+            for (itx, column) in columns.iter().enumerate() {
+                if REQUIRED_COLUMNS > itx {
+                    let to_divide = REQUIRED_COLUMNS - itx;
+                    let num_taken = min(
+                        remaining_height,
+                        (row_counter / to_divide) + usize::from(row_counter % to_divide != 0),
+                    );
+                    row_counter -= num_taken;
+                    let chunk = (&mut gauge_info).take(num_taken);
+
+                    let rows = Layout::default()
+                        .direction(Direction::Vertical)
+                        .constraints(vec![Constraint::Length(1); remaining_height])
+                        .horizontal_margin(1)
+                        .split(*column);
+
+                    for ((start_label, inner_label, ratio, style), row) in chunk.zip(rows.iter()) {
+                        f.render_widget(
+                            PipeGauge::default()
+                                .gauge_style(style)
+                                .label_style(style)
+                                .inner_label(inner_label)
+                                .start_label(start_label)
+                                .ratio(ratio)
+                                .hide_parts(hide_parts),
+                            *row,
                         );
-                        row_counter -= num_taken;
-                        let chunk = (&mut gauge_info).take(num_taken);
-
-                        let rows = Layout::default()
-                            .direction(Direction::Vertical)
-                            .constraints(vec![Constraint::Length(1); remaining_height])
-                            .horizontal_margin(1)
-                            .split(*column);
-
-                        for ((start_label, inner_label, ratio, style), row) in
-                            chunk.zip(rows.iter())
-                        {
-                            f.render_widget(
-                                PipeGauge::default()
-                                    .gauge_style(style)
-                                    .label_style(style)
-                                    .inner_label(inner_label)
-                                    .start_label(start_label)
-                                    .ratio(ratio)
-                                    .hide_parts(hide_parts),
-                                *row,
-                            );
-                        }
                     }
                 }
             }
